Read from FPL site only if fixture raw data doesn't exist

+++ create_fixture_spreadsheets.py
@@ -1,5 +1,6 @@
 import requests
 import json
+import os.path
 
 def get_data(fpl_link):
   r = requests.get(fpl_link)
@@ -10,8 +11,15 @@ def get_data(fpl_link):
     return r.content
 
 if __name__ == "__main__":
-  data = get_data('https://fantasy.premierleague.com/drf/fixtures/')
+  # https://stackoverflow.com/a/82852
+  if not os.path.isfile('raw-data/fixtures.txt'):
+    data = get_data('https://fantasy.premierleague.com/drf/fixtures/')
+    # https://stackoverflow.com/a/12309296
+    with open('raw-data/fixtures.txt', 'w') as outfile:
+      json.dump(data, outfile)
+  else:
+    #https://stackoverflow.com/a/2835672
+    with open('raw-data/fixtures.txt') as fixtures:
+      data = json.load(fixtures)
 
-  # https://stackoverflow.com/a/12309296
-  with open('fixtures.txt', 'w') as outfile:
-    json.dump(data, outfile)
+    print(data)
